fix reward bag

+++ src/items/item.cpp
@@ -471,8 +471,8 @@ Attr_ReadValue Item::readAttr(AttrTypes_t attr, PropStream& propStream)
 		}
 
 		case ATTR_WRITTENDATE: {
-			uint32_t writtenDate;
-			if (!propStream.read<uint32_t>(writtenDate)) {
+			uint64_t writtenDate;
+			if (!propStream.read<uint64_t>(writtenDate)) {
 				return ATTR_READ_ERROR;
 			}
 
@@ -858,7 +858,7 @@ void Item::serializeAttr(PropWriteStream& propWriteStream) const
 
 	if (const uint32_t writtenDate = getAttribute<uint32_t>(ItemAttribute_t::DATE)) {
 		propWriteStream.write<uint8_t>(ATTR_WRITTENDATE);
-		propWriteStream.write<uint32_t>(writtenDate);
+		propWriteStream.write<uint64_t>(writtenDate);
 	}
 
 	const std::string& writer = getString(ItemAttribute_t::WRITER);
